+++ src/video/video_png.h
@@ -25,34 +25,22 @@
  */
 #pragma once
 
-#include <QString>
-#include <QImage>
-#include <QVector>
-#include <QStringList>
-#include <boost/shared_ptr.hpp>
+#include "video.h"
 
-class AbstractVideoExport
-{
-public:
-    virtual ~AbstractVideoExport() { }
-
-    virtual QString name() const = 0;
-    virtual AbstractVideoExport * create(const unsigned int width, const unsigned int height) const = 0;
-
-    virtual void open(const QString filename) = 0;
-    virtual void close() = 0;
-    virtual void exportFrame(const QImage frame, const double s, const double t) = 0;
-};
-
-class Video
+class PngVideoExport : public AbstractVideoExport
 {
 private:
-    QVector<boost::shared_ptr<AbstractVideoExport> > exporters;
+    unsigned int width, height;
+    QString filename;
 
 public:
-    Video();
-    virtual ~Video();
+    PngVideoExport(const unsigned int width = 0, const unsigned int height = 0);
+    virtual ~PngVideoExport();
 
-    const QStringList getExporterNames();
-    AbstractVideoExport * getExporter(unsigned int idx, unsigned int width, unsigned int height);
+    virtual QString name() const;
+    virtual AbstractVideoExport * create(const unsigned int width, const unsigned int height) const;
+
+    virtual void open(const QString fileName);
+    virtual void close();
+    virtual void exportFrame(const QImage frame, const double s, const double t);
 };
